+++ NEWS.md
@@ -1,4 +1,5 @@
 # hipread 0.2.0.9000
+* progress bar will show by default even if readr hasn't been loaded (#12).
 
 # hipread 0.2.0
 * Added `yield` functions (`hipread_long_yield()` and `hipread_list_yield()`) which